refactor(api): move anonymization routes to privacy context
