+++ api/src/identity-access-management/application/user/user.admin.route.js
@@ -151,29 +151,6 @@ export const userAdminRoutes = [
   },
   {
     method: 'POST',
-    path: '/api/admin/users/{id}/anonymize',
-    config: {
-      validate: {
-        params: Joi.object({
-          id: identifiersType.userId,
-        }),
-      },
-      pre: [
-        {
-          method: (request, h) =>
-            securityPreHandlers.hasAtLeastOneAccessOf([
-              securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-              securityPreHandlers.checkAdminMemberHasRoleSupport,
-            ])(request, h),
-        },
-      ],
-      handler: (request, h) => userAdminController.anonymizeUser(request, h),
-      notes: ["- Permet à un administrateur d'anonymiser un utilisateur"],
-      tags: ['api', 'admin', 'identity-access-management', 'user'],
-    },
-  },
-  {
-    method: 'POST',
     path: '/api/admin/users/{id}/add-pix-authentication-method',
     config: {
       pre: [
