+++ api/src/privacy/application/routes.js
@@ -1,0 +1,3 @@
+import { anonymizeUserRoutes } from './anonymize-user.route.js';
+
+export const privacyRoutes = [anonymizeUserRoutes];
