+++ api/src/privacy/domain/emails/create-self-delete-user-account.email.js
