+++ api/src/privacy/domain/usecases/anonymize-user-by-itself.usecase.js
@@ -2,34 +2,27 @@ import { ForbiddenAccess } from '../../../shared/domain/errors.js';
 import { createSelfDeleteUserAccountEmail } from '../emails/create-self-delete-user-account.email.js';
 
 /**
- * @typedef {import('../../infrastructure/repositories/privacy-users-api.repository.js')} PrivacyUsersApiRepository
- */
-
-/**
  * @param{object} params
  * @param{number} params.userId
- * @param{PrivacyUsersApiRepository} privacyUsersApiRepository
  * @returns {Promise<boolean>}
  */
-export const selfDeleteUserAccount = async function ({
+export const anonymizeUserByItself = async function ({
   userId,
   locale,
   userRepository,
-  privacyUsersApiRepository,
   emailRepository,
+  anonymizeServices,
 }) {
-  const canSelfDeleteAccount = await privacyUsersApiRepository.canSelfDeleteAccount({ userId });
-
-  if (!canSelfDeleteAccount) {
-    throw new ForbiddenAccess();
-  }
+  const canAnonymize = await anonymizeServices.canAnonymizeItself({ userId });
+  if (!canAnonymize) throw new ForbiddenAccess();
 
   const user = await userRepository.get(userId);
 
   const anonymizedByUserId = userId;
   const anonymizedByUserRole = 'USER';
   const client = 'PIX_APP';
-  await privacyUsersApiRepository.anonymizeUser({ userId, anonymizedByUserId, anonymizedByUserRole, client });
+
+  await anonymizeServices.anonymizeUser({ userId, anonymizedByUserId, anonymizedByUserRole, client });
 
   if (user.email) {
     await emailRepository.sendEmailAsync(
