+++ api/tests/identity-access-management/acceptance/application/user/user.admin.route.test.js
@@ -268,73 +268,6 @@ describe('Acceptance | Identity Access Management | Application | Route | Admin 
     });
   });
 
-  describe('POST /api/admin/users/{id}/anonymize', function () {
-    let superAdmin;
-    let response;
-    let userId;
-    let certificationCenterId;
-    let organizationId;
-
-    beforeEach(async function () {
-      superAdmin = databaseBuilder.factory.buildUser.withRoleSuperAdmin();
-      userId = databaseBuilder.factory.buildUser.withRawPassword().id;
-      organizationId = databaseBuilder.factory.buildOrganization().id;
-      databaseBuilder.factory.buildMembership({
-        organizationId,
-        userId: userId,
-      });
-      certificationCenterId = databaseBuilder.factory.buildCertificationCenter().id;
-      databaseBuilder.factory.buildCertificationCenterMembership({
-        certificationCenterId,
-        userId: userId,
-      });
-      databaseBuilder.factory.buildOrganizationLearner({ userId, organizationId });
-      await databaseBuilder.commit();
-
-      response = await server.inject({
-        method: 'POST',
-        url: `/api/admin/users/${userId}/anonymize`,
-        payload: {},
-        headers: generateAuthenticatedUserRequestHeaders({ userId: superAdmin.id }),
-      });
-    });
-
-    it('anomymizes user', async function () {
-      // then
-      expect(response.statusCode).to.equal(204);
-
-      const user = await knex('users').where({ id: userId }).first();
-      expect(user.firstName).to.equal('(anonymised)');
-      expect(user.lastName).to.equal('(anonymised)');
-      expect(user.email).to.be.null;
-      expect(user.username).to.be.null;
-      expect(user.hasBeenAnonymised).to.be.true;
-    });
-
-    it('removes authentication methods', async function () {
-      // then
-      expect(response.statusCode).to.equal(204);
-
-      const authenticationMethods = await knex('authentication-methods').where({ userId });
-      expect(authenticationMethods.length).to.equal(0);
-    });
-
-    it("disables user's certification center, organization learner and organisation memberships", async function () {
-      // then
-      expect(response.statusCode).to.equal(204);
-      const certificationCenterMembership = await knex('certification-center-memberships')
-        .select()
-        .where({ certificationCenterId })
-        .first();
-      const organizationMembership = await knex('memberships').select().where({ organizationId }).first();
-      const organizationLearnerMembership = await knex('organization-learners').select().where({ organizationId });
-
-      expect(organizationMembership.disabledAt).not.to.be.null;
-      expect(certificationCenterMembership.disabledAt).not.to.be.null;
-      expect(organizationLearnerMembership.disabledAt).not.to.be.null;
-    });
-  });
-
   describe('POST /api/users/{id}/add-pix-authentication-method', function () {
     it('returns 201 HTTP status code', async function () {
       // given
