+++ api/tests/privacy/integration/domain/usecases/anonymize-user-by-admin.usecase.test.js
@@ -2,8 +2,6 @@ import sinon from 'sinon';
 
 import { RefreshToken } from '../../../../../src/identity-access-management/domain/models/RefreshToken.js';
 import { refreshTokenRepository } from '../../../../../src/identity-access-management/infrastructure/repositories/refresh-token.repository.js';
-import { LegalDocumentService } from '../../../../../src/legal-documents/domain/models/LegalDocumentService.js';
-import { LegalDocumentType } from '../../../../../src/legal-documents/domain/models/LegalDocumentType.js';
 import { usecases } from '../../../../../src/privacy/domain/usecases/index.js';
 import { PIX_ADMIN } from '../../../../../src/shared/constants.js';
 import { UserNotFoundError } from '../../../../../src/shared/domain/errors.js';
@@ -12,19 +10,11 @@ import { EMPTY_CORRELATION_INFO } from '../../../../../src/shared/infrastructure
 import { expect } from '../../../../test-helper.js';
 import { databaseBuilder, knex } from '../../../../tooling/databases.js';
 
-const { PIX_ORGA } = LegalDocumentService.VALUES;
-const { TOS } = LegalDocumentType.VALUES;
-
-describe('Integration | Privacy | Domain | UseCase | anonymize-user', function () {
-  let clock;
+describe('Integration | Privacy | Domain | UseCase | anonymize-user-by-admin', function () {
   const now = new Date('2024-04-05T03:04:05Z');
 
   beforeEach(function () {
-    clock = sinon.useFakeTimers({ now, toFake: ['Date'] });
-  });
-
-  afterEach(function () {
-    clock.restore();
+    sinon.useFakeTimers({ now, toFake: ['Date'] });
   });
 
   it(`deletes all user’s authentication methods,
@@ -33,22 +23,12 @@ describe('Integration | Privacy | Domain | UseCase | anonymize-user', function (
     disables all user’s organization memberships,
     disables all user’s certification center memberships,
     disables all user’s student prescriptions,
-    anonymizes user’s legal document acceptances,
-    anonymizes user login info,
-    anonymizes last user application connections lastLoggedAt,
-    anonymizes membership lastAccessedAt,
-    anonymizes certification center membership lastAccessedAt,
+    anonymizes user login info
     and anonymizes user`, async function () {
     // given
-
-    const user = databaseBuilder.factory.buildUser({
+    const user = databaseBuilder.factory.buildUser.withMembership({
       createdAt: new Date('2012-12-12T12:12:12Z'),
       updatedAt: new Date('2023-03-23T23:23:23Z'),
-    });
-
-    databaseBuilder.factory.buildMembership({
-      userId: user.id,
-      lastAccessedAt: new Date('2023-03-23T23:23:23Z'),
     });
 
     const admin = databaseBuilder.factory.buildUser.withRole();
@@ -56,26 +36,10 @@ describe('Integration | Privacy | Domain | UseCase | anonymize-user', function (
     const userId = user.id;
     const anonymizedByUserId = admin.id;
 
-    databaseBuilder.factory.buildCertificationCenterMembership({
-      userId,
-      lastAccessedAt: new Date('2023-03-23T23:23:23Z'),
-    });
-
-    databaseBuilder.factory.buildLastUserApplicationConnection({
-      userId,
-      application: 'orga',
-      lastLoggedAt: new Date('2023-03-23T23:23:23Z'),
-    });
+    databaseBuilder.factory.buildCertificationCenterMembership({ userId });
 
     const managingStudentsOrga = databaseBuilder.factory.buildOrganization({ isManagingStudents: true });
     databaseBuilder.factory.buildOrganizationLearner({ userId, organizationId: managingStudentsOrga.id });
-
-    const legalDocumentVersion = databaseBuilder.factory.buildLegalDocumentVersion({ service: PIX_ORGA, type: TOS });
-    databaseBuilder.factory.buildLegalDocumentVersionUserAcceptance({
-      userId: user.id,
-      legalDocumentVersionId: legalDocumentVersion.id,
-      acceptedAt: new Date('2023-03-23T23:23:23Z'),
-    }).id;
 
     const userLogin = databaseBuilder.factory.buildUserLogin({
       userId,
@@ -90,18 +54,16 @@ describe('Integration | Privacy | Domain | UseCase | anonymize-user', function (
 
     const refreshToken = RefreshToken.generate({
       userId,
+      audience: 'https://app.pix.fr',
       source: 'pix',
-      audience: 'https://app.dev.pix.fr',
       sessionId: 'random-session-id',
     });
     await refreshTokenRepository.save({ refreshToken });
 
     // when
-    await usecases.anonymizeUser({
+    await usecases.anonymizeUserByAdmin({
       userId,
-      anonymizedByUserId,
-      anonymizedByUserRole: PIX_ADMIN.ROLES.SUPER_ADMIN,
-      client: 'PIX_ADMIN',
+      updatedByUserId: anonymizedByUserId,
     });
 
     // then
@@ -128,7 +90,6 @@ describe('Integration | Privacy | Domain | UseCase | anonymize-user', function (
     expect(enabledMemberships).to.have.lengthOf(0);
     const disabledMemberships = await knex('memberships').where({ userId }).whereNotNull('disabledAt');
     expect(disabledMemberships).to.have.lengthOf(1);
-    expect(disabledMemberships[0].lastAccessedAt.toISOString()).to.equal('2023-03-01T00:00:00.000Z');
 
     const enabledCertificationCenterMemberships = await knex('certification-center-memberships')
       .where({ userId })
@@ -138,13 +99,9 @@ describe('Integration | Privacy | Domain | UseCase | anonymize-user', function (
       .where({ userId })
       .whereNotNull('disabledAt');
     expect(disabledCertificationCenterMemberships).to.have.lengthOf(1);
-    expect(disabledCertificationCenterMemberships[0].lastAccessedAt.toISOString()).to.equal('2023-03-01T00:00:00.000Z');
 
     const organizationLearners = await knex('organization-learners').where({ userId });
     expect(organizationLearners).to.have.lengthOf(0);
-
-    const userAcceptance = await knex('legal-document-version-user-acceptances').where({ userId: user.id }).first();
-    expect(userAcceptance).to.be.undefined;
 
     const anonymizedUserLogin = await knex('user-logins').where({ id: userLogin.id }).first();
     expect(anonymizedUserLogin.createdAt.toISOString()).to.equal('2012-12-01T00:00:00.000Z');
@@ -166,60 +123,9 @@ describe('Integration | Privacy | Domain | UseCase | anonymize-user', function (
     expect(anonymizedUser.lastTermsOfServiceValidatedAt).to.be.null;
     expect(anonymizedUser.lastPixCertifTermsOfServiceValidatedAt).to.be.null;
     expect(anonymizedUser.lastDataProtectionPolicySeenAt).to.be.null;
-
-    const lastUserApplicationConnection = await knex('last-user-application-connections').where({ userId }).first();
-    expect(lastUserApplicationConnection.lastLoggedAt.toISOString()).to.equal('2023-03-01T00:00:00.000Z');
   });
 
-  it('should anonymized user linked learner', async function () {
-    const user = databaseBuilder.factory.buildUser({
-      createdAt: new Date('2012-12-12T12:12:12Z'),
-      updatedAt: new Date('2023-03-23T23:23:23Z'),
-    });
-    const admin = databaseBuilder.factory.buildUser.withRole();
-    const userId = user.id;
-    const anonymizedByUserId = admin.id;
-
-    databaseBuilder.factory.buildMembership({
-      userId,
-      lastAccessedAt: new Date('2023-03-23T23:23:23Z'),
-    });
-
-    const managingStudentsOrga = databaseBuilder.factory.buildOrganization({ isManagingStudents: true });
-    const organizationLearner = databaseBuilder.factory.buildOrganizationLearner({
-      userId,
-      firstName: 'Jacqueline',
-      lastName: 'Colson',
-      email: 'jaquelinecolson@presque.fr',
-      organizationId: managingStudentsOrga.id,
-    });
-    const campaign = databaseBuilder.factory.buildCampaign({ organizationId: managingStudentsOrga.id });
-
-    databaseBuilder.factory.buildCampaignParticipation({
-      userId,
-      participantExternalId: 'jaquelinecolson',
-      campaignId: campaign.id,
-      organizationLearnerId: organizationLearner.id,
-    });
-    await databaseBuilder.commit();
-
-    // when
-    await usecases.anonymizeUser({
-      userId,
-      anonymizedByUserId,
-      anonymizedByUserRole: PIX_ADMIN.ROLES.SUPER_ADMIN,
-      client: 'PIX_ADMIN',
-    });
-
-    // then
-    const organizationLearners = await knex('organization-learners').where({ userId });
-    expect(organizationLearners).to.have.lengthOf(0);
-
-    const participations = await knex('campaign-participations').where({ userId });
-    expect(participations).lengthOf(0);
-  });
-
-  context('when anonymizedByUserId does not exist', function () {
+  context('when the admin user does not exist', function () {
     it('throws an error and does not anonymize the user', async function () {
       // given
       const user = databaseBuilder.factory.buildUser({ firstName: 'Bob' });
@@ -227,13 +133,11 @@ describe('Integration | Privacy | Domain | UseCase | anonymize-user', function (
 
       // when / then
       await expect(
-        usecases.anonymizeUser({
+        usecases.anonymizeUserByAdmin({
           userId: user.id,
-          anonymizedByUserId: 666,
-          anonymizedByUserRole: PIX_ADMIN.ROLES.SUPER_ADMIN,
-          client: 'PIX_ADMIN',
+          updatedByUserId: 666,
         }),
-      ).to.be.rejectedWith(UserNotFoundError, 'User not found for ID 666');
+      ).to.be.rejectedWith(UserNotFoundError, 'Admin not found for id: 666');
 
       const anonymizedUser = await knex('users').where({ id: user.id }).first();
       expect(anonymizedUser.hasBeenAnonymised).to.be.false;
@@ -253,11 +157,9 @@ describe('Integration | Privacy | Domain | UseCase | anonymize-user', function (
       await databaseBuilder.commit();
 
       // when
-      await usecases.anonymizeUser({
+      await usecases.anonymizeUserByAdmin({
         userId: user.id,
-        anonymizedByUserId: newAdmin.id,
-        anonymizedByUserRole: PIX_ADMIN.ROLES.SUPER_ADMIN,
-        client: 'PIX_ADMIN',
+        updatedByUserId: newAdmin.id,
       });
 
       // then
