+++ api/tests/privacy/integration/domain/usecases/anonymize-user-by-itself.usecase.test.js
@@ -1,10 +1,10 @@
-import { usecases } from '../../../../../src/identity-access-management/domain/usecases/index.js';
+import { usecases } from '../../../../../src/privacy/domain/usecases/index.js';
 import { ForbiddenAccess } from '../../../../../src/shared/domain/errors.js';
 import { featureToggles } from '../../../../../src/shared/infrastructure/feature-toggles/index.js';
 import { expect } from '../../../../test-helper.js';
 import { databaseBuilder } from '../../../../tooling/databases.js';
 
-describe('Integration | Identity Access Management | Domain | UseCase | self-delete-user-account', function () {
+describe('Integration | Privacy | Domain | UseCase | anonymize-user-by-itself', function () {
   context('when user can self delete their account', function () {
     context('when user has an email', function () {
       it('doesn’t throw ForbiddenError and creates a SendEmailJob', async function () {
@@ -13,7 +13,7 @@ describe('Integration | Identity Access Management | Domain | UseCase | self-del
         await databaseBuilder.commit();
 
         // when & then
-        await expect(usecases.selfDeleteUserAccount({ userId })).to.not.be.rejectedWith(ForbiddenAccess);
+        await expect(usecases.anonymizeUserByItself({ userId })).to.not.be.rejectedWith(ForbiddenAccess);
 
         await expect('SendEmailJob').to.have.been.performed.withJobsCount(1);
       });
@@ -26,7 +26,7 @@ describe('Integration | Identity Access Management | Domain | UseCase | self-del
         await databaseBuilder.commit();
 
         // when & then
-        await expect(usecases.selfDeleteUserAccount({ userId })).to.not.be.rejectedWith(ForbiddenAccess);
+        await expect(usecases.anonymizeUserByItself({ userId })).to.not.be.rejectedWith(ForbiddenAccess);
 
         await expect('SendEmailJob').to.have.been.performed.withJobsCount(0);
       });
@@ -42,7 +42,7 @@ describe('Integration | Identity Access Management | Domain | UseCase | self-del
       await featureToggles.set('isSelfAccountDeletionEnabled', false);
 
       // when & then
-      await expect(usecases.selfDeleteUserAccount({ userId })).to.be.rejectedWith(ForbiddenAccess);
+      await expect(usecases.anonymizeUserByItself({ userId })).to.be.rejectedWith(ForbiddenAccess);
     });
   });
 });
